reproduce tool: add support for regression testing
The reproduce tool can now reproduce previously discovered bugs.
New flags for that: `--old-bugs`, `--saved-results-file-path`.

The tool can now run tested app builds on any user-provided
samples (e.g. bugs found by other tools).
New flag for that: `--extra-bugs`.

The tool can now be run multiple times, and the result file will
have all the reproduced bugs from all these runs.

There are now two new fields in the results file for each
reproducible bug: "is_old" and "is_extra", which will be set to
true if the bug was reproduced with `--old-bugs` or `--extra-bugs`.

Also there's a new flag `--no-own-bugs` to skip reproducing bugs
in sync directory of bugbane-controlled fuzzer. This allows the
use of `--old-bugs` or `--extra-bugs` before fuzz testing.

To view examples, one can use the `-hh` flag of the tool.

Additionally, all tools now show the bugbane version

+++ bugbane/tools/reproduce/bugsamplesaver.py
@@ -76,7 +76,7 @@ class BugSampleSaver:
         Change saved sample name in input `issue_card`.
         """
         try:
-            sample_path = issue_card["sample"]
+            sample_path = str(issue_card["sample"])
             title = issue_card["title"]
         except KeyError as e:
             raise BugSampleSaverError(
@@ -85,14 +85,13 @@ class BugSampleSaver:
         save_name = self.title_to_sample_name(str(title))
         save_path = os.path.join(output_dir, save_name)
 
-        save_path = self.get_next_free_file_name(save_path)
-
-        try:
-            shutil.copyfile(str(sample_path), save_path)
-        except OSError as e:
-            raise BugSampleSaverError(
-                f"wasn't able to copy file '{sample_path}' to '{save_path}'"
-            ) from e
+        if os.path.isfile(sample_path) and not os.path.exists(save_path):
+            try:
+                shutil.copyfile(str(sample_path), save_path)
+            except OSError as e:
+                raise BugSampleSaverError(
+                    f"wasn't able to copy file '{sample_path}' to '{save_path}'"
+                ) from e
 
         issue_card["sample"] = save_name
 
@@ -108,15 +107,3 @@ class BugSampleSaver:
             normalized, new_mid_part="_-_-_", max_len=self.max_file_name_len
         )
         return shortened
-
-    @staticmethod
-    def get_next_free_file_name(wanted_path: str) -> str:
-        """
-        Generate the next available file name if `wanted_path` already exists.
-        """
-        number = 1
-        orig_wanted_path = wanted_path
-        while os.path.exists(wanted_path):
-            wanted_path = orig_wanted_path + "%04d" % number
-            number += 1
-        return wanted_path
